move components and corresponding scss file to separate folders
