+++ client/app/page.tsx
@@ -1,5 +1,5 @@
-import ChameleonGraphic from '@components/ChameleonGraphic'
-import IntroText from '@components/IntroText'
+import ChameleonGraphic from '@components/ChameleonGraphic/ChameleonGraphic'
+import IntroText from '@components/IntroText/IntroText'
 import LoginAndRegisterHandler from '@components/LoginAndRegisterHandler/LoginAndRegisterHandler'
 import styles from '@styles/App.module.scss'
