+++ client/components/LoginAndRegisterHandler/LoginAndRegisterHandler.tsx
@@ -2,7 +2,7 @@
 
 import React from 'react'
 import Image from 'next/image'
-import styles from '@styles/startSwiperAndLinks.module.scss'
+import styles from './LoginAndRegisterHandler.module.scss'
 import Link from 'next/link'
 
 function LoginAndRegisterHandler() {
